Major changes on payment process
Add views for payment to consumer and producer.

Now consumer has an option to send a direct message to the producer that contain a link that
allow producer to confirm that payment by clicking in the link.

Also producer now can check for a list of baskets requested to that cycle, confirming any
payment necessary.